% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mass_balance_optim.R
\name{back_trans}
\alias{back_trans}
\title{Back-transformation}
\usage{
back_trans(hat, years, log.trans, num.targets, season.names)
}
\arguments{
\item{hat}{A vector of estimated flow in the transformed space.}

\item{years}{A vector of all years in the study period}

\item{season.names}{A character vector containing the names of the seasons}
}
\description{
Transform the reconstructed values back to the flow space
and convert to data.table
}
