% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mass_balance_optim_old.R
\name{backtrans_old}
\alias{backtrans_old}
\title{Back-transformation}
\usage{
backtrans_old(hat, years)
}
\description{
Transform the reconstructed values back to the flow space
and convert to data.table
}
