% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R-utils.R
\name{colUnscale}
\alias{colUnscale}
\title{Unscale columns of a matrix}
\usage{
colUnscale(x, cm, csd)
}
\arguments{
\item{x}{A matrix.}

\item{cm}{A vector of column means}

\item{csd}{A vector of column standard deviations}
}
\value{
The unscaled matrix
}
\description{
Backtransform a matrix that was scaled before.
}
