% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mass-balance-optim.R
\name{cv_mb}
\alias{cv_mb}
\title{Cross-validation}
\usage{
cv_mb(
  instQ,
  pc.list,
  cv.folds,
  start.year,
  lambda = 1,
  log.trans = NULL,
  force.standardize = FALSE,
  return.type = c("mb", "metrics", "Q")
)
}
\arguments{
\item{instQ}{Instrumental data, in the same order as pc.list. The "season" column must be a factor.}

\item{pc.list}{List of PC matrices}

\item{cv.folds}{A list containing the cross validation folds}

\item{start.year}{The first year of record}

\item{lambda}{The penalty weight}

\item{log.trans}{A vector containing indices of the targets to be log-transformed. If no transformation is needed, provide \code{NULL}.}

\item{force.standardize}{If TRUE, all observations are standardized. See Details.}

\item{return.type}{If 'mb', only the objective function value is returned. If 'metrics', all metrics are returned. If 'Q', all Q predictions are returned.}
}
\description{
Cross-validation
}
\examples{
cvFolds <- make_Z(1922:2003, nRuns = 50, frac = 0.25, contiguous = TRUE)
cv <- cv_mb(p1Seasonal, pc3seasons, cvFolds, 1750, log.trans = 1:3, return.type = 'metrics')
}
