% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mass_balance_optim_old.R
\name{cv_mb_old}
\alias{cv_mb_old}
\title{Cross-validation}
\usage{
cv_mb_old(
  instQ,
  pc.dry,
  pc.wet,
  pc.ann,
  cv.folds,
  start.year,
  lambdas = 1,
  Jtype = 2,
  return.type = c("mb", "metrics", "Q")
)
}
\arguments{
\item{instQ}{Instrumental data for all seasons}

\item{return.type}{If 'mb', only the objective function value is returned. If 'metrics', all metrics are returned. If 'Q', all Q predictions are returned.}

\item{z}{Vector of the indices of the hold-out chunk}
}
\description{
Cross-validation
}
