% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mass_balance_optim_old.R
\name{cv_site_selection_old}
\alias{cv_site_selection_old}
\title{Calculate the objective function value in cross-validation}
\usage{
cv_site_selection_old(
  choice,
  pool,
  n.site.min = 5,
  seasons,
  Xpool,
  instQ,
  cv.folds,
  start.year,
  lambda = 1,
  Jtype = 2,
  use.robust.mean = FALSE
)
}
\arguments{
\item{choice}{Binary vector of choices}

\item{pool}{Data.table listing the sites with significant correlations for each season}

\item{n.site.min}{Minimum number of sites to be kept for each season}

\item{Xpool}{The input matrix of all the sites in the pool}
}
\description{
Takes the negative value for use in GA as GA maximizes. So the less negative the better.
}
