% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/density99.R
\docType{data}
\name{density99}
\alias{density99}
\title{Reexported dataset}
\format{
An object of class \code{data.frame} with 300 rows and 5 columns.
}
\usage{
density99
}
\description{
Data from the mbr package, reexported in a hacky way.
}
\keyword{datasets}
