% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_selection_functions.R
\name{input_selection}
\alias{input_selection}
\title{Input variable selection}
\usage{
input_selection(X, Y, method, nvmax, parallel = FALSE)
}
\arguments{
\item{X}{Input variables, can be a matrix or a data.frame. One variable per column. Must have column names.}

\item{Y}{Output variable}

\item{method}{Either "VSURF", "FWDselect", leaps forward", or "leaps backward".}

\item{nvmax}{Maximum number of variables; not applicable for the VSURF method.}

\item{parallel}{Only applicable if VSURF or FWDselect is used. Default = \code{FALSE}.}
}
\value{
If X has column names and \code{use.name == TRUE}, the names of the selected variables are returned, otherwise the selected column numbers are returned.
}
\description{
Four methods are in use: Variable Selection Using Random Forests (VSURF) from the \code{VSURF} package, forward stepwise selection from the \code{FWDselect} package, and forward and backward selection using branch and bound from the \code{leaps} package. The latter three use BIC as the selection criterion.
}
