% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mass_balance_optim_old.R
\name{lsq_mb_old}
\alias{lsq_mb_old}
\title{Least square with mass balance penalty}
\usage{
lsq_mb_old(hat, obs, lambdas, Jtype = 2)
}
\arguments{
\item{hat}{A vector of estimated Dry, Wet, and Annual log-transformed flow}

\item{obs}{A vector of observed Dry, Wet, and Annual log-transformed flow}

\item{Jtype}{Objective function type}

\item{lambda}{Penalty weight}
}
\value{
Objective function value: least squares plus a penalty term
}
\description{
Least square with mass balance penalty
}
