% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mass_balance_optim_old.R
\name{mb_fit_old}
\alias{mb_fit_old}
\title{Fit parameters with mass balance criterion}
\usage{
mb_fit_old(X, Y, lambdas = 1, Jtype = 2)
}
\arguments{
\item{X}{Matrix of inputs, constructed from U, V, and Z}

\item{Y}{Vector of observed log-transformed flow (following Dry, Wet, and Ann order)}

\item{lambdas}{Vector of weights to search}
}
\description{
Fit parameters with mass balance criterion
}
