% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mass_balance_optim.R
\name{mb_reconstruction}
\alias{mb_reconstruction}
\title{Mass-balance-adjusted reconstruction}
\usage{
mb_reconstruction(
  instQ,
  pc.list,
  start.year,
  lambda = 1,
  log.trans,
  num.targets
)
}
\arguments{
\item{instQ}{Instrumental data, in the same order as pc.list. The "season" column must be a factor.}

\item{pc.list}{List of PC matrices. The first element is for the first season, second element for second season, and so on. The last element is for the annual reconstruction.}

\item{start.year}{The first year of record}

\item{lambda}{The penalty weight}
}
\description{
Mass-balance-adjusted reconstruction
}
