% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mass_balance_optim_old.R
\name{mb_reconstruction_old}
\alias{mb_reconstruction_old}
\title{Mass-balance-adjusted reconstruction}
\usage{
mb_reconstruction_old(
  instQ,
  pc.dry,
  pc.wet,
  pc.ann,
  start.year,
  lambdas = 1,
  Jtype = 2
)
}
\arguments{
\item{instQ}{Instrumental data, in the same order as pc.list. The "season" column must be a factor.}

\item{pc.dry, pc.wet, pc.ann}{List of PC matrices. The first element is for the first season, second element for second season, and so on. The last element is for the annual reconstruction.}

\item{start.year}{The first year of record}

\item{lambda}{The penalty weight}
}
\description{
Mass-balance-adjusted reconstruction
}
