% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mass_balance_optim_old.R
\name{obj_fun_old}
\alias{obj_fun_old}
\title{Objective function from parameter}
\usage{
obj_fun_old(beta, X, Y, lambda, Jtype = 2)
}
\arguments{
\item{beta}{Parameters}

\item{X}{Inputs, must have columns of 1 added}

\item{Y}{Observed Dry, Wet, and Annual log-transformed flows}

\item{lambda}{Penalty weight}
}
\value{
Objective function value
}
\description{
This is a wrapper for \code{lsq_mb()}. It first calculates \code{hat}, then calls \code{lsq_mb()}.
This is used in \code{optim()}, so it returns a scalar.
}
