% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pc3seasons}
\alias{pc3seasons}
\title{Principal components of tree rings}
\format{
A list with three elements (NJ, JO, and WY), each element is a principal component matrix.
}
\source{
\url{https://www.essoar.org/doi/10.1002/essoar.10504791.1}
}
\usage{
pc3seasons
}
\description{
Principal components of the Southeast Asian Dendrochronology Network, after appropriate sites have been selected for each season.
}
\references{
Nguyen, H. T. T., Galelli, S., Xu, C., & Buckley, B. (2020). Multi-Proxy, Multi-Season Streamflow Reconstruction with Mass Balance Adjustment. Earth and Space Science Open Archive, 22. https://doi.org/10.1002/essoar.10504791.1
}
\keyword{datasets}
