% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mass-balance-optim.R
\name{prepend_ones}
\alias{prepend_ones}
\title{Prepend a column of ones}
\usage{
prepend_ones(x)
}
\arguments{
\item{x}{The input matrix}
}
\value{
x with a column of ones prepended, which is named 'Int' for 'intercept'
}
\description{
Prepend a column of ones
}
