% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reexport.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{new_expr<-}
\alias{template<-}
\alias{IC}
\alias{R2}
\alias{add_analyses}
\alias{add_models}
\alias{analyse}
\alias{analyse1}
\alias{analyse_residuals}
\alias{analyses}
\alias{as.analyses}
\alias{as.model}
\alias{as.models}
\alias{augment}
\alias{backwards}
\alias{base_model}
\alias{check_mb_analysis}
\alias{check_mb_code}
\alias{check_mb_model}
\alias{check_model_pars}
\alias{check_uniquely_named_list}
\alias{code}
\alias{coef}
\alias{coef_profile}
\alias{comment_string}
\alias{data_set}
\alias{drop_pars}
\alias{elapsed}
\alias{estimates}
\alias{fitted}
\alias{glance}
\alias{is.lmb_analysis}
\alias{is.lmb_code}
\alias{is.lmb_model}
\alias{is.mb_analyses}
\alias{is.mb_analysis}
\alias{is.mb_code}
\alias{is.mb_model}
\alias{is.mb_models}
\alias{is.mb_null_analysis}
\alias{is.syntactic}
\alias{is_bayesian}
\alias{is_frequentist}
\alias{is_namedlist}
\alias{load_model}
\alias{logLik}
\alias{log_odds_ratio2}
\alias{make_all_models}
\alias{mb_code}
\alias{mcmc_derive_data}
\alias{model}
\alias{models}
\alias{modify_data}
\alias{modify_new_data}
\alias{monitor}
\alias{new_expr}
\alias{ngens}
\alias{nmodels}
\alias{nthin}
\alias{odds_ratio2}
\alias{plot_data}
\alias{plot_residuals}
\alias{posterior_predictive_check}
\alias{predict}
\alias{proportional_change2}
\alias{proportional_difference2}
\alias{random_effects}
\alias{reanalyse}
\alias{reanalyse1}
\alias{rm_comments}
\alias{sample_size}
\alias{scalar_nlist}
\alias{sd_priors_by}
\alias{select_rescale_data}
\alias{set_analysis_mode}
\alias{simulate_residuals}
\alias{sort_by_ic}
\alias{sort_nlist}
\alias{template}
\alias{terms}
\alias{tidy}
\alias{update_model}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{embr}{\code{\link[embr]{IC}}, \code{\link[embr]{R2}}, \code{\link[embr]{add_analyses}}, \code{\link[embr]{add_models}}, \code{\link[embr]{analyse}}, \code{\link[embr]{analyse1}}, \code{\link[embr]{analyse_residuals}}, \code{\link[embr]{analyses}}, \code{\link[embr]{as.analyses}}, \code{\link[embr]{as.model}}, \code{\link[embr]{as.models}}, \code{\link[embr:reexports]{augment}}, \code{\link[embr]{backwards}}, \code{\link[embr]{base_model}}, \code{\link[embr]{check_mb_analysis}}, \code{\link[embr]{check_mb_code}}, \code{\link[embr]{check_mb_model}}, \code{\link[embr]{check_model_pars}}, \code{\link[embr]{check_uniquely_named_list}}, \code{\link[embr]{code}}, \code{\link[embr:reexports]{coef}}, \code{\link[embr]{coef_profile}}, \code{\link[embr]{comment_string}}, \code{\link[embr]{data_set}}, \code{\link[embr]{drop_pars}}, \code{\link[embr]{elapsed}}, \code{\link[embr:reexports]{estimates}}, \code{\link[embr:reexports]{fitted}}, \code{\link[embr:reexports]{glance}}, \code{\link[embr]{is.lmb_analysis}}, \code{\link[embr]{is.lmb_code}}, \code{\link[embr]{is.lmb_model}}, \code{\link[embr]{is.mb_analyses}}, \code{\link[embr]{is.mb_analysis}}, \code{\link[embr]{is.mb_code}}, \code{\link[embr]{is.mb_model}}, \code{\link[embr]{is.mb_models}}, \code{\link[embr]{is.mb_null_analysis}}, \code{\link[embr]{is.syntactic}}, \code{\link[embr]{is_bayesian}}, \code{\link[embr]{is_frequentist}}, \code{\link[embr]{is_namedlist}}, \code{\link[embr]{load_model}}, \code{\link[embr:reexports]{logLik}}, \code{\link[embr:reexports]{log_odds_ratio2}}, \code{\link[embr]{make_all_models}}, \code{\link[embr]{mb_code}}, \code{\link[embr:reexports]{mcmc_derive_data}}, \code{\link[embr]{model}}, \code{\link[embr]{models}}, \code{\link[embr]{modify_data}}, \code{\link[embr]{modify_new_data}}, \code{\link[embr]{monitor}}, \code{\link[embr]{new_expr}}, \code{\link[embr:new_expr-set]{new_expr<-}}, \code{\link[embr]{ngens}}, \code{\link[embr]{nmodels}}, \code{\link[embr]{nthin}}, \code{\link[embr:reexports]{odds_ratio2}}, \code{\link[embr]{plot_data}}, \code{\link[embr]{plot_residuals}}, \code{\link[embr]{posterior_predictive_check}}, \code{\link[embr:reexports]{predict}}, \code{\link[embr:reexports]{proportional_change2}}, \code{\link[embr:reexports]{proportional_difference2}}, \code{\link[embr]{random_effects}}, \code{\link[embr]{reanalyse}}, \code{\link[embr]{reanalyse1}}, \code{\link[embr]{rm_comments}}, \code{\link[embr]{sample_size}}, \code{\link[embr]{scalar_nlist}}, \code{\link[embr]{sd_priors_by}}, \code{\link[embr]{select_rescale_data}}, \code{\link[embr]{set_analysis_mode}}, \code{\link[embr]{simulate_residuals}}, \code{\link[embr]{sort_by_ic}}, \code{\link[embr]{sort_nlist}}, \code{\link[embr]{template}}, \code{\link[embr:template-set]{template<-}}, \code{\link[embr:reexports]{terms}}, \code{\link[embr:reexports]{tidy}}, \code{\link[embr]{update_model}}}
}}

