% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R-utils.R
\name{rowUnscale}
\alias{rowUnscale}
\title{Unscale rows of a matrix}
\usage{
rowUnscale(x, rm, rsd)
}
\arguments{
\item{x}{A matrix.}

\item{rm}{A vector of row means}

\item{rsd}{A vector of row standard deviations}
}
\value{
The unscaled matrix
}
\description{
Backtransform a matrix that was scaled before.
}
