% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_selection_functions.R
\name{wPCA}
\alias{wPCA}
\title{Weighted PCA based on correlation, a la Cook et al (2010). Each column of X will be weighted by an exponent r^p, i.e.
\eqn{W_i = X_i*r_i^p}}
\usage{
wPCA(
  X,
  rho = 1,
  p = 0,
  min.variance = 0.95,
  use.eigen = TRUE,
  return.matrix = FALSE
)
}
\arguments{
\item{X}{A matrix of variables}

\item{rho}{Vector of correlations between each variable and the output, ignored if \code{rho == 1}}

\item{p}{Weight exponent}

\item{min.variance}{Minimum fraction of variance explained by the retained PCs}

\item{use.eigen}{IF TRUE, only PCs whose eigenvalue > 1 are kept.}

\item{return.matrix}{If \code{TRUE}, a matrix is returned, otherwise a \code{data.table} is returned.}
}
\description{
Weighted PCA based on correlation, a la Cook et al (2010). Each column of X will be weighted by an exponent r^p, i.e.
\eqn{W_i = X_i*r_i^p}
}
