% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mass_balance_optim.R
\name{cv_mb}
\alias{cv_mb}
\title{Cross-validation}
\usage{
cv_mb(
  instQ,
  pc.list,
  cv.folds,
  start.year,
  lambda = 1,
  log.trans,
  num.targets,
  return.type = c("mb", "metrics", "Q")
)
}
\arguments{
\item{instQ}{Instrumental data, in the same order as pc.list. The "season" column must be a factor.}

\item{pc.list}{List of PC matrices. The first element is for the first season, second element for second season, and so on. The last element is for the annual reconstruction.}

\item{cv.folds}{A list containing the cross validation folds}

\item{start.year}{The first year of record}

\item{lambda}{The penalty weight}

\item{return.type}{If 'mb', only the objective function value is returned. If 'metrics', all metrics are returned. If 'Q', all Q predictions are returned.}

\item{pc.list.inst}{List of PC matrices}
}
\description{
Cross-validation
}
