% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mbr-package.R
\docType{package}
\name{mbr}
\alias{mbr}
\alias{mbr-package}
\title{mbr: Mass Balance Reconstruction}
\description{
Mass balance reconstruction
    See paper
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/ntthung/mbr}
  \item Report bugs at \url{https://github.com/ntthung/mbr/issues}
}

}
\keyword{internal}
